+++ R/add_NE_layer.R
@@ -1,7 +1,7 @@
 #' @noRd
 add_ne_layer <- function(layer_name, scale, extent_projected) {
 
-  geometry <- featurecla <- scalerank <- type <- NULL
+  geometry <- featurecla <- scalerank <- type <- geom <- NULL
 
   temp_ne_dir <- tools::R_user_dir("rnaturalearth", "cache")
 
@@ -60,14 +60,18 @@ add_ne_layer <- function(layer_name, scale, extent_projected) {
       if (grepl("the file .* seems not to exist", e, ignore.case = TRUE) ||
           grepl("Failed to download", e, ignore.case = TRUE) ||
           grepl("HTTP status was 404", e, ignore.case = TRUE)) {
-        message(paste0("Attempting to download '", layer_name,
-                       "' data due to previous load error."))
+        message(paste0("Attempting to download '", layer_name, "' data."))
         rnaturalearth::ne_download(scale = scale,
                                    returnclass = "sf",
                                    type = layer_name,
                                    category = category,
                                    load = FALSE,
                                    destdir = temp_ne_dir)
+        rnaturalearth::ne_load(scale = scale,
+                               returnclass = "sf",
+                               type = layer_name,
+                               category = category,
+                               temp_ne_dir)
       } else {
         stop(e) # Re-throw other unhandled errors
       }
@@ -79,21 +83,21 @@ add_ne_layer <- function(layer_name, scale, extent_projected) {
 
   }
 
-  extent_projected <- sf::st_transform(sf::st_as_sfc(extent_projected),
-                                       crs = "ESRI:54012")
+  # extent_projected <- sf::st_transform(sf::st_as_sfc(extent_projected),
+  #                                      crs = "ESRI:54012")
 
   # Attempt to perform cropping for efficiency FIRST
   # If it fails, validate first
   processed_layer <- tryCatch({
     layer_raw %>%
-      sf::st_transform(crs = "ESRI:54012") %>%
+  #    sf::st_transform(crs = "ESRI:54012") %>%
       dplyr::group_by(scalerank, featurecla) %>%
       dplyr::reframe(geometry = sf::st_crop(geometry, extent_projected)) %>%
       sf::st_as_sf()
   }, error = function(e) {
     layer_raw %>%
       sf::st_make_valid() %>%
-      sf::st_transform(crs = "ESRI:54012") %>%
+   #   sf::st_transform(crs = "ESRI:54012") %>%
       dplyr::group_by(scalerank, featurecla) %>%
       dplyr::reframe(geometry = sf::st_crop(geometry, extent_projected)) %>%
       sf::st_as_sf()
@@ -101,7 +105,7 @@ add_ne_layer <- function(layer_name, scale, extent_projected) {
 
   # Then validate and filter (on the now smaller dataset)
   processed_layer <- processed_layer %>%
-    sf::st_make_valid() %>%
+   # sf::st_make_valid() %>%
     dplyr::filter(!sf::st_is_empty(geometry)) # Filter out empty geometries
 
   return(processed_layer)