+++ R/compute_indicator_workflow.R
@@ -24,37 +24,45 @@
 #'     rarefaction).
 #'   * 'hill2': Hill-Simpson diversity (estimated by coverage-based
 #'     rarefaction).
-#' @param dim_type (Optional) Dimension to calculate indicator over (time: 'ts',
-#'  or space: 'map')
+#' @param dim_type (Optional) Dimension to calculate indicator over time: 'ts',
+#'  or space: 'map'. (Default: 'map')
 #' @param ci_type (Optional) Type of bootstrap confidence intervals to
-#'  calculate. (Default: "norm". Select "none" to avoid calculating bootstrap
-#'  CIs.)
-#' @param cell_size (Optional) Length of grid cell sides, in km. (Default: 10
-#'  for country, 100 for continent or world)
+#'  calculate. (Default: "norm"). Select "none" to avoid calculating bootstrap
+#'  CIs.
+#' @param cell_size (Optional) Length of grid cell sides, in km or degrees.
+#'  If NULL, this will be automatically determined according to the geographical
+#'  level selected. This is 100 km or 1 degree for 'continent' or 'world', 10 km
+#'  or (for a degree-based CRS) the native resolution of the cube for 'country',
+#'  'sovereignty' or 'geounit'. If level is set to 'cube', cell size will be the
+#'  native resolution of the cube for a degree-based CRS, or for a km-based CRS,
+#'  the cell size will be determined by the area of the cube: 100 km for cubes
+#'  larger than 1 million sq km, 10 km for cubes between 10 thousand and 1
+#'  million sq km, 1 km for cubes between 100 and 10 thousand sq km, and 0.1 km
+#'  for cubes smaller than 100 sq km. (Default: NULL)
 #' @param level (Optional) Spatial level: 'cube', 'continent', 'country',
 #'  'world', 'sovereignty', or 'geounit'. (Default: 'cube')
-#' @param region (Optional) The region of interest (e.g., "Europe"). (Default:
-#'  "Europe")
+#' @param region (Optional) The region of interest (e.g., "Europe"). This
+#'  parameter is ignored if level is set to 'cube' or 'world'. (Default: NULL)
 #' @param ne_type (Optional) The type of Natural Earth data to download:
-#'  'countries', 'map_units', 'sovereignty', or 'tiny_countries'. (Default:
-#'  "countries")
+#'  'countries', 'map_units', 'sovereignty', or 'tiny_countries'. This parameter
+#'  is ignored if level is set to 'cube' or 'world'. (Default: "countries")
 #' @param ne_scale (Optional) The scale of Natural Earth data to download:
 #'  'small' - 110m, 'medium' - 50m, or 'large' - 10m. (Default: "medium")
 #' @param output_crs (Optional) The CRS you want for your calculated indicator.
 #'  (Leave blank to let the function choose a default based on grid reference
-#'  system)
-#' @param first_year (Optinal) Exclude data before this year. (Uses all data in
+#'  system.)
+#' @param first_year (Optional) Exclude data before this year. (Uses all data in
 #'  the cube by default.)
 #' @param last_year (Optional) Exclude data after this year. (Uses all data in
 #'  the cube by default.)
 #' @param spherical_geometry (Optional) If set to FALSE, will temporarily
 #'  disable spherical geometry while the function runs. Should only be used to
-#'  solve specific issues. (Default is TRUE)
-#' @param make_valid (Optional) Calls st_make_valid() from the sf package.
-#'  Increases processing time but may help if you are getting polygon errors.
-#'  (Default is FALSE).
+#'  solve specific issues. (Default is TRUE).
+#' @param make_valid (Optional) Calls st_make_valid() from the sf package
+#'  after creating the grid. Increases processing time but may help if you are
+#'  getting polygon errors. (Default is FALSE).
 #' @param num_bootstrap (Optional) Set the number of bootstraps to calculate for
-#'  generating confidence intervals. (Default: 1000)
+#'  generating confidence intervals. (Default: 100)
 #' @param shapefile_path (optional) Path of an external shapefile to merge into
 #'  the workflow. For example, if you want to calculate your indicator
 #'  particular features such as protected areas or wetlands.
@@ -64,12 +72,25 @@
 #'  NOT a .wkt the CRS will be determined automatically.
 #' @param invert (optional) Calculate an indicator over the inverse of the
 #'  shapefile (e.g. if you have a protected areas shapefile this would calculate
-#'  an indicator over all non protected areas)
+#'  an indicator over all non protected areas within your cube). Default is
+#'  FALSE.
+#' @param include_land (Optional) Include occurrences which fall within the
+#'  land area. Default is TRUE. *Note that this purely a geographic filter, and
+#'  does not filter based on whether the occurrence is actually terrestrial.
+#'  Grid cells which fall partially on land and partially on ocean will be
+#'  included even if include_land is FALSE. To exclude terrestrial and/or
+#'  freshwater taxa, you must manually filter your data cube before calculating
+#'  your indicator.
 #' @param include_ocean (Optional) Include occurrences which fall outside the
 #'  land area. Default is TRUE. Set as "buffered_coast" to include a set buffer
-#'  size around the land area rather than the entire ocean area.
+#'  size around the land area rather than the entire ocean area. *Note that this
+#'  is purely a geographic filter, and does not filter based on whether the
+#'  occurrence is actually marine. Grid cells which fall partially on land and
+#'  partially on ocean will be included even if include_ocean is FALSE. To
+#'  exclude marine taxa, you must manually filter your data cube before
+#'  calculating your indicator.
 #' @param buffer_dist_km (Optional) The distance to buffer around the land if
-#'  include_ocean is set to "buffered_coast".
+#'  include_ocean is set to "buffered_coast". Default is 50 km.
 #' @param force_grid (Optional) Forces the calculation of a grid even if this
 #'  would not normally be part of the pipeline, e.g. for time series. This
 #'  setting is required for the calculation of rarity, and is turned on by the
@@ -121,6 +142,7 @@ compute_indicator_workflow <- function(data,
                                        shapefile_path = NULL,
                                        shapefile_crs = NULL,
                                        invert = FALSE,
+                                       include_land = TRUE,
                                        include_ocean = TRUE,
                                        buffer_dist_km = 50,
                                        force_grid = FALSE,
@@ -138,6 +160,11 @@ compute_indicator_workflow <- function(data,
   # Check that obs column exists
   if (!"obs" %in% colnames(data$data)) {
     stop("No occurrences found in the data.")
+  }
+
+  # Check that user has not excluded both land and ocean data
+  if (!include_land && !include_ocean) {
+    stop("You must include either land or ocean data, or both.")
   }
 
   # List of indicators that require grid cells for temporal calculations
@@ -246,8 +273,13 @@ compute_indicator_workflow <- function(data,
   } else {
     coord_range
   }
-  region <- ifelse(level == "cube", "cube",
-                   ifelse(level == "world", "world", region))
+  region <- if (level == "cube") {
+    "cube"
+  } else if (level == "world") {
+    "world"
+  } else {
+    region
+  }
 
   # Check shapefile path and load if found
   if (!is.null(shapefile_path)) {
@@ -305,9 +337,9 @@ compute_indicator_workflow <- function(data,
 
     # Choose appropriate projection CRS
     if (data$grid_type == "eea") {
-      projected_crs <- "EPSG:3035"
+      projected_crs <- cube_crs
     } else if (data$grid_type == "mgrs") {
-      projected_crs <- guess_utm_epsg(data)
+      projected_crs <- cube_crs
     } else if (data$grid_type == "eqdgc") {
       projected_crs <- guess_utm_epsg(cube_bbox_latlong)
     } else {
@@ -475,19 +507,18 @@ compute_indicator_workflow <- function(data,
     }
 
     # Retrieve and validate Natural Earth data
-    map_data <- get_ne_data(projected_crs,
-                            bbox_latlong,
-                            region,
-                            level,
-                            ne_type,
-                            ne_scale,
-                            include_ocean,
-                            buffer_dist_km) %>%
-      sf::st_make_valid()
+    map_data_list <- get_ne_data(projected_crs,
+                                 bbox_latlong,
+                                 region,
+                                 level,
+                                 ne_type,
+                                 ne_scale,
+                                 include_land,
+                                 include_ocean,
+                                 buffer_dist_km)
+    map_data <- map_data_list$combined
+    saved_layer <- map_data_list$saved
 
-    if (data$grid_type == "eea") {
-      map_data <- sf::st_transform(map_data, crs = projected_crs)
-    }
   }
 
   if (dim_type == "map" || force_grid == TRUE) {
@@ -500,52 +531,35 @@ compute_indicator_workflow <- function(data,
 
     sf::st_agr(grid) <- "constant"
 
-    # The following intersection operation requires special error handling
-    # because it fails when the grid contains invalid geometries.
-    # Therefore when invalid geometries are encountered, it will retry the
-    # operation with spherical geometry turned off. This often succeeds.
-    result <- NULL  # Initialize to capture result of intersection
-    tryCatch({
-      # Attempt without altering the spherical geometry setting
-      result <- grid %>%
-        sf::st_intersection(map_data) %>%
-        dplyr::select(dplyr::all_of(c("cellid", "geometry")),
-                      dplyr::any_of("area"))
-    }, error = function(e) {
-      if (grepl("Error in wk_handle.wk_wkb", e) ||
-          grepl("TopologyException", e)) {
-        message(
-          paste0(
-            "Encountered a geometry error during intersection. This may be ",
-            "due to invalid polygons in the grid."
-          )
+    # Define the object to be used for intersection
+    # If a shapefile is provided, use it directly. Otherwise, use the map data.
+    if (!is.null(shapefile)) {
+      if (invert) {
+        # If invert is TRUE, we want the area *outside* the shapefile but inside
+        # the cube. We use st_difference for this.
+        intersection_target <- sf::st_difference(
+          cube_polygon_projected, sf::st_union(shapefile_projected)
         )
       } else {
-        stop(e)
+        # If invert is FALSE, we simply want the area of the shapefile
+        intersection_target <- shapefile_projected
       }
-    })
-    if (is.null(result)) {
-      # If intersection failed, turn off spherical geometry
-      message("Retrying the intersection with spherical geometry turned off.")
-      sf::sf_use_s2(FALSE)
-      # Retry the intersection operation
-      result <- grid %>%
-        sf::st_intersection(map_data) %>%
-        dplyr::select(dplyr::all_of(c("cellid", "geometry")),
-                      dplyr::any_of("area"))
-      # Notify success after retry
-      message("Intersection succeeded with spherical geometry turned off.")
+
+      intersection_target <- sf::st_as_sf(intersection_target)
+    } else {
+      # No shapefile provided, so we intersect with the Natural Earth data
+      intersection_target <- map_data
     }
+
+    sf::st_agr(intersection_target) <- "constant"
+
+    # Intersect grid with intersection target
+    clipped_grid <- intersect_grid_with_polygon(grid, intersection_target)
+
     if (spherical_geometry == TRUE) {
       # Restore original spherical setting
       sf::sf_use_s2(original_s2_setting)
     }
-    # Check if there is any spatial intersection
-    if (nrow(result) == 0) {
-      stop("No spatial intersection between map data and grid.")
-    }
-    # Set grid to result
-    clipped_grid <- result
 
     # Assign data to grid
     data_final <- sf::st_join(data_projected, clipped_grid)
@@ -581,6 +595,10 @@ compute_indicator_workflow <- function(data,
     diversity_grid <-
       clipped_grid %>%
       dplyr::left_join(indicator, by = "cellid")
+
+    # Get bbox of original grid before transformation
+    original_bbox <- intersect_grid_with_polygon(grid, saved_layer) %>%
+      sf::st_union()
 
     # Transform to output CRS
     diversity_grid <- sf::st_transform(diversity_grid, crs = output_crs)
@@ -630,6 +648,7 @@ compute_indicator_workflow <- function(data,
                                        cell_size_units = output_units,
                                        map_level = level,
                                        map_region = region,
+                                       map_type = ne_type,
                                        kingdoms = kingdoms,
                                        num_families = num_families,
                                        num_species = num_species,
@@ -637,7 +656,8 @@ compute_indicator_workflow <- function(data,
                                        last_year = last_year,
                                        num_years = num_years,
                                        species_names = species_names,
-                                       years_with_obs = years_with_obs)
+                                       years_with_obs = years_with_obs,
+                                       original_bbox = original_bbox)
   } else {
 
     # Build indicator_ts object
@@ -645,6 +665,7 @@ compute_indicator_workflow <- function(data,
                                       div_type = type,
                                       map_level = level,
                                       map_region = region,
+                                      map_type = ne_type,
                                       kingdoms = kingdoms,
                                       num_families = num_families,
                                       num_species = num_species,