+++ R/prepare_map_for_plot.R
@@ -1,5 +1,13 @@
 #' @noRd
-prepare_map_for_plot <- function(x, xlims, ylims, layers, scale, crop_to_grid) {
+prepare_map_for_plot <- function(x,
+                                 xlims,
+                                 ylims,
+                                 layers,
+                                 scale,
+                                 crop_to_grid,
+                                 crop_by_region,
+                                 projection,
+                                 map_expansion_factor) {
 
   # Check that object to plot is the correct class
   if (!inherits(x, "indicator_map")) {
@@ -39,19 +47,25 @@ prepare_map_for_plot <- function(x, xlims, ylims, layers, scale, crop_to_grid) {
 
   # Prepare map data with correct layers, CRS, bounding box, etc.
   map_data_list <- prepare_map_data(data = x$data,
-                                    projection = x$projection,
+                                    projection = projection,
                                     map_lims = map_lims,
                                     xlims = xlims,
                                     ylims = ylims,
                                     map_data_sf = map_data_sf,
                                     layers = layers,
                                     scale = scale,
-                                    crop_to_grid = crop_to_grid
+                                    crop_to_grid = crop_to_grid,
+                                    map_region = x$map_region,
+                                    map_level = x$map_level,
+                                    map_type = x$map_type,
+                                    crop_by_region = crop_by_region,
+                                    map_expansion_factor = map_expansion_factor
   )
 
   # Get map data and layer list
   map_surround <- map_data_list$map_surround
   layer_list <- map_data_list$layer_list
+  map_lims <- map_data_list$map_lims
 
   # Return map limits, surrounding map data, and layer list
   return(list(